+++ previous-versions/sync-for-expenses-version-1/pkg/models/operations/initiatesync.go
@@ -30,14 +30,10 @@ func (o *InitiateSyncRequest) GetCompanyID() string {
 type InitiateSyncResponse struct {
 	// HTTP response content type for this operation
 	ContentType string
-	// If model is incorrect
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
 	RawResponse *http.Response
-	// Returns the newly created SyncId
-	SyncInitiated *shared.SyncInitiated
 }
 
 func (o *InitiateSyncResponse) GetContentType() string {
@@ -45,13 +41,6 @@ func (o *InitiateSyncResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *InitiateSyncResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *InitiateSyncResponse) GetStatusCode() int {
@@ -66,11 +55,4 @@ func (o *InitiateSyncResponse) GetRawResponse() *http.Response {
 		return nil
 	}
 	return o.RawResponse
-}
-
-func (o *InitiateSyncResponse) GetSyncInitiated() *shared.SyncInitiated {
-	if o == nil {
-		return nil
-	}
-	return o.SyncInitiated
 }