+++ previous-versions/sync-for-expenses-version-1/pkg/models/operations/savecompanyconfiguration.go
@@ -32,8 +32,6 @@ type SaveCompanyConfigurationResponse struct {
 	CompanyConfiguration *shared.CompanyConfiguration
 	// HTTP response content type for this operation
 	ContentType string
-	// The request made is not valid.
-	ErrorMessage *shared.ErrorMessage
 	// HTTP response status code for this operation
 	StatusCode int
 	// Raw HTTP response; suitable for custom response parsing
@@ -52,13 +50,6 @@ func (o *SaveCompanyConfigurationResponse) GetContentType() string {
 		return ""
 	}
 	return o.ContentType
-}
-
-func (o *SaveCompanyConfigurationResponse) GetErrorMessage() *shared.ErrorMessage {
-	if o == nil {
-		return nil
-	}
-	return o.ErrorMessage
 }
 
 func (o *SaveCompanyConfigurationResponse) GetStatusCode() int {